feat: premium tier visibility + shared models dir
Surface the exact premium fair-use classification ($5+/M input or
$15+/M output, as enforced by the gateway) everywhere users need it:

- apps/ui /models: Pricing Tier filter (premium/standard), gem icon
  on premium rows, models annotated server-side via isPremiumModel
- apps/ui /models/premium: new SEO category page listing exactly the
  premium models, linked from the directory, sitemap and docs
- apps/code /models: full models directory (same search, filters,
  pagination as apps/ui) via the AllModels component extracted to
  @llmgateway/shared/components, with DevPass chrome and a premium
  vs standard explainer
- apps/code dashboard: weekly premium allowance progress bar (same
  UsageBar as plan credits) with days+hours reset countdown, backed
  by new devPlanPremium* fields on GET /dev-plans/status

The models directory (all-models, model-card, category filters, API
types/fetchers) now lives in packages/shared with footer/CTA/link
base injected per app; apps/ui keeps thin shims so existing imports
are unchanged.
