+++ apps/code/src/app/sitemap.ts
@@ -19,6 +19,12 @@ export default function sitemap(): MetadataRoute.Sitemap {
 			priority: 0.8,
 		},
 		{
+			url: `${baseUrl}/models`,
+			lastModified: new Date(),
+			changeFrequency: "weekly",
+			priority: 0.8,
+		},
+		{
 			url: `${baseUrl}/pricing`,
 			lastModified: new Date(),
 			changeFrequency: "weekly",
