+++ apps/code/src/components/DevPassModelsDirectory.tsx
@@ -1,0 +1,42 @@
+"use client";
+
+import { ArrowRight } from "lucide-react";
+import Link from "next/link";
+
+import { Footer } from "@/components/Footer";
+import { Button } from "@/components/ui/button";
+
+import { AllModels as SharedAllModels } from "@llmgateway/shared/components";
+
+import type { ComponentProps } from "react";
+
+type SharedAllModelsProps = ComponentProps<typeof SharedAllModels>;
+
+export function DevPassModelsDirectory({
+	uiUrl,
+	...rest
+}: Omit<SharedAllModelsProps, "footer" | "renderCta" | "modelHrefBase"> & {
+	uiUrl: string;
+}) {
+	return (
+		<SharedAllModels
+			{...rest}
+			modelHrefBase={uiUrl}
+			footer={<Footer />}
+			renderCta={({ size, className, iconClassName, onClick }) => (
+				<Button
+					variant="default"
+					size={size}
+					className={className}
+					onClick={onClick}
+					asChild
+				>
+					<Link href="/signup">
+						Get DevPass
+						<ArrowRight className={iconClassName} />
+					</Link>
+				</Button>
+			)}
+		/>
+	);
+}
