+++ apps/docs/content/learn/model-categories.mdx
@@ -21,7 +21,7 @@ Every model in the gateway is sorted into a category. Categories power dashboard
 | **Premium**  | High-cost frontier / flagship models — priced at **$15+ per million output tokens** or **$5+ per million input tokens** |
 | **Standard** | Every other model — the broad catalog of fast, cost-effective everyday models                                           |
 
-You can browse the full catalog on the [**Supported Models**](https://llmgateway.io/models) page and filter by use case, capabilities, provider, price, and context size.
+The classification is computed automatically from catalogue prices — there is no hand-picked list. To see exactly which models are premium right now, check the live [**Premium Models**](https://llmgateway.io/models/premium) page, or browse the full catalog on the [**Supported Models**](https://llmgateway.io/models) page and filter by pricing tier, use case, capabilities, provider, price, and context size.
 
 ## Fair-use caps on premium models (DevPass only)
