+++ apps/ui/src/app/models/premium/page.tsx
@@ -1,0 +1,10 @@
+import {
+	buildCategoryMetadata,
+	ModelCategoryPage,
+} from "@/components/models/category-page";
+
+export const metadata = buildCategoryMetadata("premium");
+
+export default function PremiumModelsPage() {
+	return <ModelCategoryPage slug="premium" />;
+}
