+++ apps/ui/src/lib/fetch-models.ts
@@ -1,119 +1,24 @@
 import { cache } from "react";
 
+import {
+	fetchModelsFromApi,
+	fetchProvidersFromApi,
+} from "@llmgateway/shared/components";
+
 import { getConfig } from "./config-server";
 
 import type { DiscountData } from "./discount";
+import type { ApiModel, ApiProvider } from "@llmgateway/shared/components";
 
-export interface ApiProvider {
-	id: string;
-	createdAt: string;
-	name: string | null;
-	description: string | null;
-	streaming: boolean | null;
-	cancellation: boolean | null;
-	color: string | null;
-	website: string | null;
-	announcement: string | null;
-	serviceTiers?: Array<{
-		id: string;
-		name: string;
-		multiplier: number;
-		description?: string;
-	}> | null;
-	status: "active" | "inactive";
-}
-
-export interface ApiModelProviderMapping {
-	id: string;
-	createdAt: string;
-	modelId: string;
-	providerId: string;
-	externalId: string;
-	region?: string | null;
-	inputPrice: string | null;
-	outputPrice: string | null;
-	cachedInputPrice: string | null;
-	cacheWriteInputPrice: string | null;
-	cacheWriteInputPrice1h: string | null;
-	imageInputPrice: string | null;
-	imageOutputPrice: string | null;
-	imageInputTokensByResolution: Record<string, number> | null;
-	imageOutputTokensByResolution: Record<string, number> | null;
-	inputCharacterPrice: string | null;
-	outputAudioPrice: string | null;
-	requestPrice: string | null;
-	ocrPagePrice?: string | null;
-	contextSize: number | null;
-	maxOutput: number | null;
-	quantization?: string | null;
-	streaming: boolean;
-	vision: boolean | null;
-	reasoning: boolean | null;
-	reasoningEfforts?:
-		| ("none" | "minimal" | "low" | "medium" | "high" | "xhigh" | "max")[]
-		| null;
-	reasoningOutput: string | null;
-	reasoningMaxTokens: boolean | null;
-	tools: boolean | null;
-	jsonOutput: boolean | null;
-	jsonOutputSchema: boolean | null;
-	webSearch: boolean | null;
-	webSearchPrice: string | null;
-	supportedVideoSizes: string[] | null;
-	supportedVideoDurationsSeconds: number[] | null;
-	supportsVideoAudio: boolean | null;
-	supportsVideoWithoutAudio: boolean | null;
-	perSecondPrice: Record<string, string> | null;
-	pricingTiers: Array<{
-		name: string;
-		upToTokens: number | null;
-		inputPrice: string;
-		outputPrice: string;
-		cachedInputPrice: string | null;
-		cacheReadInputPrice: string | null;
-		cacheWriteInputPrice: string | null;
-		cacheWriteInputPrice1h: string | null;
-	}> | null;
-	serviceTiers?: string[] | null;
-	discount: string | null;
-	stability: "stable" | "beta" | "unstable" | "experimental" | null;
-	supportedParameters: string[] | null;
-	deprecatedAt: string | null;
-	deactivatedAt: string | null;
-	status: "active" | "inactive";
-}
-
-export interface ApiModel {
-	id: string;
-	createdAt: string;
-	releasedAt: string | null;
-	name: string | null;
-	aliases: string[] | null;
-	description: string | null;
-	family: string;
-	free: boolean | null;
-	output: string[] | null;
-	stability: "stable" | "beta" | "unstable" | "experimental" | null;
-	status: "active" | "inactive";
-	mappings: ApiModelProviderMapping[];
-}
+export type {
+	ApiModel,
+	ApiModelProviderMapping,
+	ApiProvider,
+} from "@llmgateway/shared/components";
 
 export const fetchModels = cache(async (): Promise<ApiModel[]> => {
 	const config = getConfig();
-	try {
-		const response = await fetch(`${config.apiBackendUrl}/internal/models`, {
-			next: { revalidate: 60 },
-		});
-		if (!response.ok) {
-			console.error("Failed to fetch models:", response.statusText);
-			return [];
-		}
-		const data = await response.json();
-		return data.models ?? [];
-	} catch (error) {
-		console.error("Error fetching models:", error);
-		return [];
-	}
+	return await fetchModelsFromApi(config.apiBackendUrl);
 });
 
 export const fetchModelDiscounts = cache(
@@ -139,18 +44,5 @@ export const fetchModelDiscounts = cache(
 
 export const fetchProviders = cache(async (): Promise<ApiProvider[]> => {
 	const config = getConfig();
-	try {
-		const response = await fetch(`${config.apiBackendUrl}/internal/providers`, {
-			next: { revalidate: 60 },
-		});
-		if (!response.ok) {
-			console.error("Failed to fetch providers:", response.statusText);
-			return [];
-		}
-		const data = await response.json();
-		return data.providers ?? [];
-	} catch (error) {
-		console.error("Error fetching providers:", error);
-		return [];
-	}
+	return await fetchProvidersFromApi(config.apiBackendUrl);
 });
