+++ packages/shared/package.json
@@ -68,6 +68,7 @@
 		"@radix-ui/react-slot": "1.2.4",
 		"@radix-ui/react-switch": "1.2.6",
 		"@radix-ui/react-tabs": "^1.1.13",
+		"@radix-ui/react-toggle": "1.1.10",
 		"@radix-ui/react-tooltip": "^1.2.8",
 		"@radix-ui/react-use-controllable-state": "1.2.2",
 		"class-variance-authority": "0.7.1",
@@ -77,8 +78,10 @@
 		"decimal.js": "10.5.0",
 		"embla-carousel-react": "8.6.0",
 		"lucide-react": "0.561.0",
+		"next": "16.2.10",
 		"next-themes": "^0.4.6",
 		"pretty-bytes": "7.1.0",
+		"prism-react-renderer": "2.4.1",
 		"react": "19.2.7",
 		"react-dom": "19.2.7",
 		"react-hook-form": "7.65.0",
