+++ packages/shared/src/components/index.tsx
@@ -2,6 +2,12 @@ export * from "./integration-guides-grid";
 export * from "./integration-icons";
 export * from "./log-card";
 export * from "./model-selector";
+export * from "./models-directory/all-models";
+export * from "./models-directory/api-types";
+export * from "./models-directory/model-card";
+export * from "./models-directory/model-category-filters";
+export * from "./models-directory/model-code-example-dialog";
+export * from "./models-directory/model-status-badge";
 export * from "./multi-model-selector";
 export * from "./multi-provider-selector";
 export * from "./provider-icons";
