+++ packages/shared/src/components/ui/index.tsx
@@ -2,6 +2,7 @@ export * from "./alert";
 export * from "./avatar";
 export * from "./badge";
 export * from "./button";
+export * from "./card";
 export * from "./carousel";
 export * from "./checkbox";
 export * from "./collapsible";
@@ -25,6 +26,8 @@ export * from "./sidebar";
 export * from "./skeleton";
 export * from "./sonner";
 export * from "./switch";
+export * from "./table";
 export * from "./tabs";
 export * from "./textarea";
+export * from "./toggle";
 export * from "./tooltip";
